chore : dto,entity 그외 테스트
dto 와 entity date 형식 수정
펫 관련부분 테스트후 수정 완료

+++ src/main/java/com/mogumogu/moru/pet/dto/PetInfoDTO.java
@@ -17,8 +17,8 @@ public class PetInfoDTO {
     private UserInfoDTO userInfoDTO;
     private PetBreedDTO petBreedDTO;
     private String piName;
-    private Integer piAdult;
-    private Integer piNeutering;
+    private byte piAdult;
+    private byte piNeutering;
     public static PetInfoDTO toDTO(PetInfo petInfo){
         return  PetInfoDTO.builder()
                 .piNum(petInfo.getPiNum())
